pyproject.toml
[tool.poetry]
name = "ELF_RSS"
version = "2.5.4"
description = "QQ机器人 RSS订阅 插件，订阅源建议选择 RSSHub"
authors = ["Quan666 <i@oy.mk>"]
license = "GPL-3.0-only"
readme = "README.md"
packages = [
    { include = "ELF_RSS2", from = "src/plugins" },
]
homepage = "https://github.com/Quan666/ELF_RSS"
repository = "https://github.com/Quan666/ELF_RSS"
keywords = ["nonebot", "nonebot2", "rss" ,"elf" ,"rsshub"]

[tool.poetry.dependencies]
python = "^3.8.3"
aiohttp = "^3.8.1"
arrow = "^1.2.0"
bbcode = "^1.1.0"
emoji = "^0.5.4"
feedparser = "^6.0.0"
deep-translator = "^1.7.0"
ImageHash = "^4.2.0"
magneturi = "^1.3"
nonebot-adapter-onebot = "^2.0.0b1"
nonebot-plugin-apscheduler = "^0.1.2"
nonebot-plugin-guild-patch = "^0.1.1"
nonebot2 = "^2.0.0b1"
Pillow = "^9.0.0"
pydantic = "^1.9.0"
pyquery = "^1.4.3"
python-qbittorrent = "^0.4.2"
tenacity = "^7.0.0"
tinydb = "^4.6.1"
typing-extensions = "^3.10.0.2"
yarl = "^1.7.2"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
